refactor proposal narrative action routes
